+++ tests/test_ratelimits.py
@@ -3,7 +3,7 @@ A test that checks the user request limit control function.
 """
 import re
 import pytest
-from users.ratelimits import RateLimiter
+from users import RateLimiter
 
 
 @pytest.mark.order(6)
